Explicit cast from int to u8
It is needed to pass "/Wx" switch when using MSVC

+++ RGFW.h
@@ -7444,7 +7444,7 @@ RGFW_event* RGFW_window_checkEvent(RGFW_window* win) {
 			if (msg.message == WM_XBUTTONDOWN)
 				win->event.button = RGFW_mouseMisc1 + (GET_XBUTTON_WPARAM(msg.wParam) == XBUTTON2);
 			else win->event.button = (msg.message == WM_LBUTTONDOWN) ? RGFW_mouseLeft :
-									 (msg.message == WM_RBUTTONDOWN) ? RGFW_mouseRight : RGFW_mouseMiddle;
+									 (msg.message == WM_RBUTTONDOWN) ? (u8)RGFW_mouseRight : (u8)RGFW_mouseMiddle;
 
 			win->event.type = RGFW_mouseButtonPressed;
 			_RGFW->mouseButtons[win->event.button].prev = _RGFW->mouseButtons[win->event.button].current;
@@ -7455,7 +7455,7 @@ RGFW_event* RGFW_window_checkEvent(RGFW_window* win) {
 			if (msg.message == WM_XBUTTONUP)
 				win->event.button = RGFW_mouseMisc1 + (GET_XBUTTON_WPARAM(msg.wParam) == XBUTTON2);
 			else win->event.button = (msg.message == WM_LBUTTONUP) ? RGFW_mouseLeft :
-									 (msg.message == WM_RBUTTONUP) ? RGFW_mouseRight : RGFW_mouseMiddle;
+									 (msg.message == WM_RBUTTONUP) ? (u8)RGFW_mouseRight : (u8)RGFW_mouseMiddle;
 			win->event.type = RGFW_mouseButtonReleased;
 			_RGFW->mouseButtons[win->event.button].prev = _RGFW->mouseButtons[win->event.button].current;
 			_RGFW->mouseButtons[win->event.button].current = 0;
